+++ packages/plugins/1password/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @varlock/1password-plugin
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies [[`4d436ff`](https://github.com/dmno-dev/varlock/commit/4d436ff42863136fb5ebb7016e525ef54732ea20), [`ca51993`](https://github.com/dmno-dev/varlock/commit/ca5199371cd6126794e215f67cfcc5f20342eaaa)]:
+  - varlock@0.5.0
+
 ## 0.2.2
 
 ### Patch Changes